Fixup graphs and charts not displaying on main page
Using PHP8.1, the main page was not displaying the graphs and charts, just circling with a wait cursor.
The Apache errorlog here /var/log/apache2/error.log
was showing this error...

...PHP Fatal error:
   Method ezcGraphCoordinate::__set_state() must be static in
        /var/www/html/crashfix/protected/vendors/ezcomponents/Graph/src/structs/coordinate.php on line 58

There is a later GitHub repo for the Graph components, with this commit...
https://github.com/zetacomponents/Graph/commit/a9e534018300be4b42f0c766ea7aea7ec0a084c8

It updates three php files altering the __set_state functions.
Updating these three files with this commit was enough to restore the main page to a working state.